Add adhoc query editor
