+++ src/views/ConfigEditor/constants.ts
@@ -1,0 +1,6 @@
+export const COMPRESSION_TYPE_OPTIONS = [
+  { label: 'gzip', value: 'gzip' },
+  { label: 'br', value: 'br' },
+  { label: 'deflate', value: 'deflate' },
+  { label: 'zstd', value: 'zstd' },
+]
