+++ src/views/QueryEditor/components/QueryTextEditor/QueryTextEditor.tsx
@@ -32,8 +32,6 @@ const FORMAT_OPTIONS = [
 
 export const QueryTextEditor = ({
   query,
-  height,
-  onEditorMount,
   onSqlChange,
   onFieldChange,
   formattedData,
@@ -81,10 +79,8 @@ export const QueryTextEditor = ({
     <>
       <SQLCodeEditor
         datasource={datasource}
-        height={height}
         onSqlChange={onSqlChange}
         query={query}
-        onEditorMount={onEditorMount}
         onRunQuery={onRunQuery}
       />
       {!areAdHocFiltersAvailable && (
@@ -95,11 +91,7 @@ export const QueryTextEditor = ({
             onFieldChange({
               fieldName: 'adHocFilters',
               value: tagsList.map((item: string) => {
-                const [
-                  key,
-                  operator,
-                  value
-                ] = item.split(' ');
+                const [key, operator, value] = item.split(' ');
 
                 return { key, operator, value };
               }),
